requirements.txt
paho-mqtt>=1.6.0,<2.0.0
requests>=2.28.0
PyYAML>=6.0
